spotify-telegram-bot/pyproject.toml
[tool.poetry]
name = "spotify-telegram-bot"
version = "0.1.0"
description = "A Telegram bot for searching and sharing Spotify content."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.7"
spotipy = "^2.19.0"
dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"